pyproject.toml
[project]
name = "spineai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.9.1",
    "aioboto3>=15.0.0",
    "aiohttp>=3.12.13",
    "aiosmtplib>=4.0.1",
    "azure-ai-documentintelligence>=1.0.2",
    "bcrypt>=4.3.0",
    "celery>=5.5.3",
    "fastapi>=0.115.14",
    "faster-whisper>=1.1.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain-community>=0.3.27",
    "openai>=1.93.0",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pydicom>=3.0.1",
    "pyjwt>=2.10.1",
    "pymupdf>=1.26.3",
    "python-gdcm>=3.0.26",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "ruff>=0.12.2",
    "stripe>=12.3.0",
    "tiktoken>=0.9.0",
    "tomlkit>=0.13.3",
    "tortoise-orm[asyncpg]>=0.25.1",
    "tortoise-vector>=0.1.4",
    "uvicorn>=0.35.0",
    "uvicorn-worker>=0.3.0",
    "whisper>=1.1.10",
]

[tool.aerich]
tortoise_orm = "app.db.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
